# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carefy api",
        "contact": {
            "email": "devabnerugeda@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/pacientes": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "colete todas as informações",
                "description": "Coletar todos os dados de todos os pacientes",
                "operationId": "67a2883311115dd5978caee65a63e646",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "500": {
                        "description": "Erro no sistema"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "",
                "description": "Crie seu paciente",
                "operationId": "8acfc295905f210a96af583640d5183c",
                "requestBody": {
                    "description": "Insira corretamente todas as informações",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Codigo_Paciente": {
                                        "type": "integer",
                                        "example": "1234"
                                    },
                                    "Nome": {
                                        "type": "string",
                                        "example": "Abner Ugeda"
                                    },
                                    "CEP": {
                                        "type": "string",
                                        "example": "00000-000"
                                    },
                                    "Logradouro": {
                                        "type": "string",
                                        "example": "Rua floriano"
                                    },
                                    "Bairro": {
                                        "type": "string",
                                        "example": "Rua floriano"
                                    },
                                    "numero_casa": {
                                        "type": "integer",
                                        "example": "123"
                                    },
                                    "complemento": {
                                        "type": "string",
                                        "example": "apto 71"
                                    },
                                    "UF": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "Data_Nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "Telefone": {
                                        "type": "string",
                                        "example": "(16) 99792-6503"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso! Usuário foi cadastrado"
                    },
                    "500": {
                        "description": "Erro no sistema"
                    },
                    "200": {
                        "description": "Código do paciente já cadastrado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Informações inválidas"
                    }
                }
            }
        },
        "/pacientes/{id}": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Colete a informação de apenas um Paciente",
                "description": "Coletar todos os dados de um paciente pelo seu id",
                "operationId": "13f72d945300ca36d6aa14f2a0373a0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "202": {
                        "description": "Paciente não encontrado"
                    },
                    "500": {
                        "description": "Erro no sistema"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "",
                "description": "Edite o seu paciente",
                "operationId": "e48086743d6bf17ed405ebbfce87176f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "request": "UpdatePacienteRequest",
                    "description": "Insira corretamente todas as informações",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Codigo_Paciente": {
                                        "type": "integer",
                                        "example": "1234"
                                    },
                                    "Nome": {
                                        "type": "string",
                                        "example": "Abner Ugeda"
                                    },
                                    "CEP": {
                                        "type": "string",
                                        "example": "00000-000"
                                    },
                                    "Logradouro": {
                                        "type": "string",
                                        "example": "Rua floriano"
                                    },
                                    "Bairro": {
                                        "type": "string",
                                        "example": "Rua floriano"
                                    },
                                    "numero_casa": {
                                        "type": "integer",
                                        "example": "123"
                                    },
                                    "complemento": {
                                        "type": "string",
                                        "example": "apto 71"
                                    },
                                    "UF": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "Data_Nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "Telefone": {
                                        "type": "string",
                                        "example": "(16) 99792-6503"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso! Usuário foi cadastrado"
                    },
                    "500": {
                        "description": "Erro no sistema"
                    },
                    "202": {
                        "description": "Paciente não encontrado"
                    },
                    "404": {
                        "description": "Informações inválidas"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Colete a informação de apenas um Paciente",
                "description": "Coletar todos os dados de todos os pacientes",
                "operationId": "d8876ee67937bc8fb44837f474b23a03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "202": {
                        "description": "Paciente não encontrado"
                    },
                    "500": {
                        "description": "Erro no sistema"
                    }
                }
            }
        },
        "/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "colete todas as informações",
                "description": "Coletar todos os dados de todos as Tags",
                "operationId": "a411cff9cc7ba5ef475fb340fe358295",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "500": {
                        "description": "Erro no sistema"
                    }
                }
            }
        },
        "/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Colete a informação de apenas uma Tag",
                "description": "Coletar os dados de uma Tag pelo seu id",
                "operationId": "d21879d535a01e762ff8f5d6d40a56d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "202": {
                        "description": "Tag não encontrada"
                    },
                    "500": {
                        "description": "Erro no sistema"
                    }
                }
            }
        },
        "/tags/{codigo}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Colete a informação de apenas uma Tag pelo Código de Tag",
                "description": "Coletar os dados de uma tag pelo seu código",
                "operationId": "429b8221140ddda1101c7d103272adae",
                "parameters": [
                    {
                        "name": "codigo",
                        "in": "path",
                        "description": "Buscar por codigo da tag",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "202": {
                        "description": "Tag não encontrada"
                    },
                    "500": {
                        "description": "Erro no sistema"
                    }
                }
            }
        }
    }
}